requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
cachelib==0.10.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Session==0.4.0
Flask-SSE==1.0.0
frozenlist==1.3.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==6.0.4
numpy==1.21.3
openai==0.27.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.5.1
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.15
Werkzeug==2.0.2
yarl==1.8.2
python-decouple==3.8
loguru~=0.6.0
mysql-connector-python~=8.0.32
tiktoken~=0.3.2
ping3~=4.0.4